Add event listeners to sidebar navigation links
When a link is hovered over, its icon will adopt a new class
in order to change its color.

Additionally, the inner wrapper containing the icon and label will adopt
a new class in order to scale these elements instead of the entire link.

When no longer hovered over, the classes are removed.